requirements.txt
vk-api==11.9.9
python-dotenv==1.0.0
asyncio
aiohttp==3.9.1 